# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_LanreAkintayo_outpost_internal_dto.CreateEventTypeRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        type: string
    required:
    - name
    type: object
  github_com_LanreAkintayo_outpost_internal_dto.CreateSubscriptionRequest:
    properties:
      event_type_id:
        type: string
    required:
    - event_type_id
    type: object
...
  github_com_LanreAkintayo_outpost_internal_dto.EventResponse:
    properties:
      application_id:
        type: string
      created_at:
        type: string
      event_type_id:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      payload:
        items:
          type: integer
        type: array
      recipient_id:
        type: string
    type: object
...
  github_com_LanreAkintayo_outpost_internal_dto.EventTypeResponse:
    properties:
      application_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_LanreAkintayo_outpost_internal_dto.EventTypeSummary:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  github_com_LanreAkintayo_outpost_internal_dto.IngestEventResponse:
    properties:
      created_at:
        type: string
      event_type:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      queued_deliveries:
        type: integer
      recipient_id:
        type: string
      status:
        type: string
    type: object
...
  github_com_LanreAkintayo_outpost_internal_dto.SendEventRequest:
    properties:
      event_type:
        maxLength: 255
        minLength: 1
        type: string
      idempotency_key:
        type: string
      payload:
        items:
          type: integer
        type: array
      recipient_id:
        type: string
    required:
    - event_type
    - payload
    type: object
...
  github_com_LanreAkintayo_outpost_internal_dto.SubscriptionResponse:
    properties:
      created_at:
        type: string
      endpoint_id:
        type: string
      event_type:
        $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.EventTypeSummary'
      event_type_id:
        type: string
      id:
        type: string
    type: object
...
  github_com_LanreAkintayo_outpost_internal_response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact:
    name: Outpost Support
  description: High-performance, fault-tolerant webhook delivery platform with exponential
    retries, rate limiting, HMAC signing, and dead-letter queues.
  license:
    name: MIT
  title: Outpost Webhook Delivery Engine API
  version: "1.0"
paths:
...
  /api/v1/endpoints/{id}/subscriptions:
    get:
      description: List all active event type subscriptions for a specific endpoint
      parameters:
      - description: Endpoint UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.SubscriptionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List endpoint subscriptions
      tags:
      - Subscriptions
    post:
      consumes:
      - application/json
      description: Subscribe an endpoint to receive events of a specific event type
      parameters:
      - description: Endpoint UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Subscribe endpoint
      tags:
      - Subscriptions
  /api/v1/endpoints/{id}/subscriptions/{event_type_id}:
    delete:
      description: Remove a subscription binding between an endpoint and an event
        type
      parameters:
      - description: Endpoint UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Event Type UUID
        format: uuid
        in: path
        name: event_type_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unsubscribe endpoint
      tags:
      - Subscriptions
  /api/v1/event-types:
    get:
      description: Fetch all event types registered in the authenticated application
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.EventTypeResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List event types
      tags:
      - Event Types
    post:
      consumes:
      - application/json
      description: Define a new event type name in dot-notation (e.g. payment.completed)
      parameters:
      - description: Event type details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.CreateEventTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.EventTypeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create event type
      tags:
      - Event Types
  /api/v1/event-types/{id}:
    delete:
      description: Delete an event type by UUID
      parameters:
      - description: Event Type UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete event type
      tags:
      - Event Types
    get:
      description: Retrieve single event type details by UUID
      parameters:
      - description: Event Type UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.EventTypeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get event type by ID
      tags:
      - Event Types
  /api/v1/events:
    post:
      consumes:
      - application/json
      description: Ingest a webhook event, check idempotency, schedule delivery attempts
        to subscribed endpoints, and enqueue tasks
      parameters:
      - description: Event payload and routing parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.SendEventRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.IngestEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ingest event
      tags:
      - Events
  /api/v1/events/{id}:
    get:
      description: Fetch details of an ingested event by UUID
      parameters:
      - description: Event UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get event by ID
      tags:
      - Events
...
securityDefinitions:
  BearerAuth:
    description: Enter your API key with the Bearer prefix, e.g. 'Bearer op_live_...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"